exaone api 구성 완료
